requirements.txt
asgiref==3.8.1
azure-core==1.30.2
azure-storage-blob==12.22.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
isodate==0.6.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
